[CI] Run angular linting and tests, fix current issues, upload test artifacts (#144)

+++ servers/dsub/jobs/test/test_jobs_controller_local.py
@@ -35,7 +35,8 @@ class TestJobsControllerLocal(BaseTestCases.JobsControllerTestCase):
         os.mkdir(self.log_path)
 
     def tearDown(self):
-        shutil.rmtree(self.testing_root)
+        if os.environ.get('KEEP_TEST_DSUB_FILES') != 'true':
+            shutil.rmtree(self.testing_root)
         tempfile.tempdir = None
 
     def create_app(self):
